fix package_days and heading with null

+++ app/Http/Controllers/Admin/AdminProfileController.php
@@ -39,7 +39,9 @@ class AdminProfileController extends Controller
                 'photo' => 'image|mimes:jpg,jpeg,png,gif',
             ]);
 
-            unlink(public_path('uploads/' . $admin_data->photo));
+            if(file_exists(public_path('uploads/' . $admin_data->photo))) {
+                unlink(public_path('uploads/' . $admin_data->photo));
+            }
 
             $ext = $request->file('photo')->extension();
             $final_name = 'admin' . '.' . $ext;
